consolidated data pipe files into one
